requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0
scikit-learn>=0.24.0
torch>=1.9.0
transformers>=4.11.0
openai>=0.27.0
tqdm>=4.62.0
plotly>=5.3.0
latex2mathml>=3.61.0
python-dotenv>=0.19.0

# LLM inference (choose based on your setup)
accelerate>=0.24.0

# For Unsloth (if using)
unsloth[colab-new]>=2024.1

# Data analysis
scipy>=1.11.0

# Logging and utilities
tqdm>=4.66.0 